# api-specs/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080"}],"paths":{"/vehicle-types/{id}":{"get":{"tags":["vehicle-type-controller"],"operationId":"getVehicleTypeFlat","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["vehicle-type-controller"],"operationId":"updateVehicleTypeParameter","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/vehicle-models/{id}":{"get":{"tags":["vehicle-type-model-controller"],"operationId":"getVehicleTypeFlat_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["vehicle-type-model-controller"],"operationId":"updateVehicleTypeModel","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"color","in":"query","required":true,"schema":{"type":"string"}},{"name":"length","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VehicleTypeModel"}}}}}}},"/menu/{menuId}":{"get":{"tags":["menu-controller"],"operationId":"getMenuById","parameters":[{"name":"menuId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MenuResponse"}}}}}},"put":{"tags":["menu-controller"],"operationId":"updateMenuById","parameters":[{"name":"menuId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MenuResponse"}}}}}},"delete":{"tags":["menu-controller"],"operationId":"deleteMenuById","parameters":[{"name":"menuId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/menu/list":{"get":{"tags":["menu-controller"],"operationId":"getMenus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuResponse"}}}}}}},"put":{"tags":["menu-controller"],"operationId":"updateMenus","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuResponse"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuResponse"}}}}}}},"post":{"tags":["menu-controller"],"operationId":"createMenus","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuRequest"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuResponse"}}}}}}}},"/vehicle/vehicle-route/{versionId}":{"post":{"tags":["vehicle-controller"],"operationId":"getVehicleRoute","parameters":[{"name":"versionId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/vehicle/generate-vehicle-route/{versionId}":{"post":{"tags":["vehicle-controller"],"operationId":"generateVehicleRoute","parameters":[{"name":"versionId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/vehicle-types":{"get":{"tags":["vehicle-type-controller"],"operationId":"getAllVehicleTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VehicleType"}}}}}}},"post":{"tags":["vehicle-type-controller"],"operationId":"createVehicleTypeWithParameters","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/vehicle-types/delete":{"post":{"tags":["vehicle-type-controller"],"operationId":"deleteVehicleTypeParameters","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"integer","format":"int64"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/vehicle-models":{"get":{"tags":["vehicle-type-model-controller"],"operationId":"getAllVehicleTypeModel","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VehicleTypeModel"}}}}}}},"post":{"tags":["vehicle-type-model-controller"],"operationId":"createVehicleModel","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"color","in":"query","required":true,"schema":{"type":"string"}},{"name":"length","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/vehicle-models/delete":{"post":{"tags":["vehicle-type-model-controller"],"operationId":"deleteVehicleTypeModels","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"integer","format":"int64"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/signal/{versionId}":{"get":{"tags":["signal-controller"],"operationId":"getSignal","parameters":[{"name":"versionId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignalNodeResponseData"}}}}}},"post":{"tags":["signal-controller"],"operationId":"saveSignal","parameters":[{"name":"versionId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignalSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/schema/{layer-key}":{"get":{"tags":["schema-controller"],"operationId":"getSchemaByLayerKey","parameters":[{"name":"layer-key","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LayerSchemaResponse"}}}}}},"post":{"tags":["schema-controller"],"operationId":"updateSchemata","parameters":[{"name":"layer-key","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SchemaFieldsRequest"}}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/pavement-marking/{versionId}":{"get":{"tags":["pavement-marking-controller"],"operationId":"getPavementMarkingByVersion","parameters":[{"name":"versionId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoadAssetData"}}}}}},"post":{"tags":["pavement-marking-controller"],"operationId":"savePavementMarking","parameters":[{"name":"versionId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PavementMarkingSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/menu":{"post":{"tags":["menu-controller"],"operationId":"createMenu","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MenuResponse"}}}}}}},"/signal/histories/{versionId}":{"get":{"tags":["signal-controller"],"operationId":"getLogsByVersion","parameters":[{"name":"versionId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SignalLogs"}}}}}}}},"/schema":{"get":{"tags":["schema-controller"],"operationId":"getSchemata","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LayerSchemaResponse"}}}}}}}},"/scenario":{"get":{"tags":["scenario-controller"],"operationId":"getAllScenarios","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Scenario"}}}}}}}},"/scenario/{id}":{"get":{"tags":["scenario-controller"],"operationId":"getScenarioById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Scenario"}}}}}}},"/scenario/{id}/versions":{"get":{"tags":["scenario-controller"],"operationId":"getScenarioVersions","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScenarioVersion"}}}}}}}},"/public-transit/station/rail/{scenarioKey}":{"get":{"tags":["rail-station-controller"],"operationId":"getRailStationsByScenarioKey","parameters":[{"name":"scenarioKey","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RailPublicTransitResponse"}}}}}}},"/public-transit/station/bus/{scenarioKey}":{"get":{"tags":["bus-station-controller"],"operationId":"getBusStationsByScenarioKey","parameters":[{"name":"scenarioKey","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PublicTransitResponse"}}}}}}},"/pavement-marking/histories/{versionId}":{"get":{"tags":["pavement-marking-controller"],"operationId":"getLogsByVersion_1","parameters":[{"name":"versionId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PavementMarkingLogs"}}}}}}}},"/network/{scenarioKey}":{"get":{"tags":["network-controller"],"operationId":"getNetworkByScenarioKey","parameters":[{"name":"scenarioKey","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NetworkResponse"}}}}}}},"/menu/tree":{"get":{"tags":["menu-controller"],"operationId":"getMenuTree","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuTreeResponse"}}}}}}}},"/menu/depth/{depth}":{"get":{"tags":["menu-controller"],"operationId":"getMenusByDepth","parameters":[{"name":"depth","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuResponse"}}}}}}}},"/menu/code/{menuCode}":{"get":{"tags":["menu-controller"],"operationId":"getMenuByCode","parameters":[{"name":"menuCode","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MenuResponse"}}}}}}},"/layer/list":{"get":{"tags":["layer-controller"],"operationId":"getAllLayers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LayerResponse"}}}}}}}},"/layer/group":{"get":{"tags":["layer-controller"],"operationId":"getAllLayerGroup","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LayerGroupResponse"}}}}}}}},"/layer/group/{groupKey}":{"get":{"tags":["layer-controller"],"operationId":"getLayerGroup","parameters":[{"name":"groupKey","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LayerGroupResponse"}}}}}}}},"components":{"schemas":{"VehicleTypeModel":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"color":{"type":"string"},"fileName":{"type":"string"},"filePath":{"type":"string"},"length":{"type":"string"}}},"MenuRequest":{"type":"object","properties":{"menuId":{"type":"integer","format":"int64"},"menuCode":{"type":"string"},"language":{"type":"string"},"nameKor":{"type":"string"},"nameEn":{"type":"string"},"parentsId":{"type":"integer","format":"int64"},"rootId":{"type":"integer","format":"int64"},"depth":{"type":"integer","format":"int32"},"sortOrder":{"type":"integer","format":"int32"},"available":{"type":"string"},"accessRole":{"type":"string","enum":["ROLE_ADMIN","ROLE_USER"]}}},"MenuResponse":{"type":"object","properties":{"menuId":{"type":"integer","format":"int64"},"menuCode":{"type":"string"},"language":{"type":"string"},"nameKor":{"type":"string"},"nameEn":{"type":"string"},"parentsId":{"type":"integer","format":"int64"},"rootId":{"type":"integer","format":"int64"},"depth":{"type":"integer","format":"int32"},"sortOrder":{"type":"integer","format":"int32"},"available":{"type":"string"},"accessRole":{"type":"string","enum":["ROLE_ADMIN","ROLE_USER"]}}},"Cartesian3":{"type":"object","properties":{"x":{"type":"number","format":"double"},"y":{"type":"number","format":"double"},"z":{"type":"number","format":"double"}}},"Polyline":{"type":"object","properties":{"positions":{"type":"array","items":{"$ref":"#/components/schemas/Cartesian3"}}}},"Road":{"type":"object","properties":{"linkId":{"type":"string"},"laneId":{"type":"string"},"polyline":{"$ref":"#/components/schemas/Polyline"},"baseEasting":{"type":"number","format":"double"},"baseNorthing":{"type":"number","format":"double"},"targetEasting":{"type":"number","format":"double"},"targetNorthing":{"type":"number","format":"double"}}},"RoadResponse":{"type":"object","properties":{"roads":{"type":"array","items":{"$ref":"#/components/schemas/Road"}}}},"VehicleRequest":{"type":"object","properties":{"numVehicle":{"type":"integer","format":"int32"},"roadEntities":{"$ref":"#/components/schemas/RoadResponse"},"speedFactor":{"type":"integer","format":"int32"}}},"Detail":{"type":"object","properties":{"field":{"type":"string"},"newValue":{},"oldValue":{},"guid":{"type":"string"},"timestamp":{"type":"string"}}},"LogsData":{"type":"object","properties":{"added":{"type":"array","items":{"$ref":"#/components/schemas/Detail"}},"deleted":{"type":"array","items":{"$ref":"#/components/schemas/Detail"}},"modified":{"type":"array","items":{"$ref":"#/components/schemas/Detail"}}}},"SignalResponse":{"type":"object","properties":{"nodeId":{"type":"string"},"turnId":{"type":"string"},"turning":{"type":"string"},"type":{"type":"string"},"connectionId":{"type":"string"},"get__guid":{"type":"string"}}},"SignalSaveRequest":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/SignalResponse"}},"logs":{"$ref":"#/components/schemas/LogsData"}}},"CreateFieldOptionRequest":{"type":"object","properties":{"value":{"type":"string"}}},"CreateFieldRequest":{"type":"object","properties":{"name":{"type":"string"},"nullable":{"type":"boolean"},"defaultValue":{"type":"string"},"readOnly":{"type":"boolean"},"status":{"type":"string","enum":["ACTIVE","INACTIVE","DELETED"]},"inputType":{"type":"string"},"options":{"type":"array","items":{"$ref":"#/components/schemas/CreateFieldOptionRequest"}}}},"SchemaFieldsRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fieldsToCreate":{"type":"array","items":{"$ref":"#/components/schemas/CreateFieldRequest"}},"fieldsToUpdate":{"type":"array","items":{"$ref":"#/components/schemas/UpdateFieldRequest"}},"fieldIdsToDelete":{"type":"array","items":{"type":"integer","format":"int64"}}}},"UpdateFieldOption":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"value":{"type":"string"}}},"UpdateFieldRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"nullable":{"type":"boolean"},"defaultValue":{"type":"string"},"readOnly":{"type":"boolean"},"status":{"type":"string","enum":["ACTIVE","INACTIVE","DELETED"]},"inputType":{"type":"string"},"options":{"type":"array","items":{"$ref":"#/components/schemas/UpdateFieldOption"}},"optionsToCreate":{"type":"array","items":{"$ref":"#/components/schemas/CreateFieldOptionRequest"}},"optionIdsToDelete":{"type":"array","items":{"type":"integer","format":"int64"}}}},"Coordinates":{"type":"object","properties":{"lat":{"type":"number","format":"double"},"lng":{"type":"number","format":"double"}}},"PavementMarkingData":{"type":"object","properties":{"id":{"type":"string"},"angle":{"type":"number","format":"double"},"cellId":{"type":"integer","format":"int32"},"offset":{"type":"number","format":"double"},"laneRef":{"type":"integer","format":"int32"},"linkRef":{"type":"integer","format":"int32"},"markingType":{"type":"string"},"coordinates":{"type":"array","items":{"$ref":"#/components/schemas/Coordinates"}}}},"PavementMarkingSaveRequest":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/PavementMarkingData"}},"logs":{"$ref":"#/components/schemas/LogsData"}}},"VehicleType":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"vehicleId":{"type":"string"},"name":{"type":"string"},"v2x":{"type":"string"},"drt":{"type":"string"},"maxPax":{"type":"string"}}},"SignalNodeResponseData":{"type":"object","properties":{"signals":{"type":"array","items":{"$ref":"#/components/schemas/SignalResponse"}}}},"SignalLogs":{"type":"object","properties":{"versionId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"data":{"$ref":"#/components/schemas/LogsData"}}},"ColumnOption":{"type":"object","properties":{"value":{"type":"string"}}},"LayerSchemaFieldResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"inputType":{"type":"string"},"defaultValue":{"type":"string"},"readOnly":{"type":"boolean"},"nullable":{"type":"boolean"},"status":{"type":"string","enum":["ACTIVE","INACTIVE","DELETED"]},"options":{"type":"array","items":{"$ref":"#/components/schemas/LayerSchemaOptionResponse"}}}},"LayerSchemaOptionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"value":{"type":"string"}}},"LayerSchemaResponse":{"type":"object","properties":{"layerId":{"type":"integer","format":"int64"},"layerName":{"type":"string"},"schemaColumns":{"type":"array","items":{"$ref":"#/components/schemas/SchemaColumn"}},"schemata":{"type":"array","items":{"$ref":"#/components/schemas/Schema"}}}},"Schema":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"status":{"type":"string"},"fields":{"type":"array","items":{"$ref":"#/components/schemas/LayerSchemaFieldResponse"}}}},"SchemaColumn":{"type":"object","properties":{"configKey":{"type":"string"},"inputType":{"type":"string"},"options":{"type":"array","items":{"$ref":"#/components/schemas/ColumnOption"}}}},"Scenario":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"key":{"type":"string"},"longitude":{"type":"number","format":"double"},"latitude":{"type":"number","format":"double"},"label":{"type":"string"},"description":{"type":"string"},"versions":{"type":"array","items":{"$ref":"#/components/schemas/ScenarioVersion"}}}},"ScenarioVersion":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"key":{"type":"string"},"label":{"type":"string"},"insertDate":{"type":"string","format":"date-time"},"modifyDate":{"type":"string","format":"date-time"}}},"ExitResponse":{"type":"object","properties":{"id":{"type":"string"},"linkRef":{"type":"string"},"offset":{"type":"number","format":"double"},"accessTime":{"type":"string"},"coord":{"type":"string"},"coordinates":{"$ref":"#/components/schemas/Coordinates"}}},"RailPublicTransitResponse":{"type":"object","properties":{"railStations":{"type":"array","items":{"$ref":"#/components/schemas/RailStationResponse"}}}},"RailStationResponse":{"type":"object","properties":{"id":{"type":"string"},"transitMode":{"type":"string","enum":["subway","bus"]},"lineList":{"type":"string"},"type":{"type":"string","enum":["island","side","face_to_face","staggered"]},"address":{"type":"string"},"center":{"type":"string"},"exits":{"type":"array","items":{"$ref":"#/components/schemas/ExitResponse"}},"timetables":{"type":"array","items":{"$ref":"#/components/schemas/TimetableResponse"}},"coordinates":{"$ref":"#/components/schemas/Coordinates"}}},"TimetableResponse":{"type":"object","properties":{"dayOfWeek":{"type":"string"},"lineId":{"type":"string"},"times":{"type":"array","items":{"type":"string"}}}},"BusLineResponse":{"type":"object","properties":{"list":{"type":"string"}}},"BusStationResponse":{"type":"object","properties":{"id":{"type":"string"},"transitMode":{"type":"string","enum":["subway","bus"]},"linkRef":{"type":"integer","format":"int32"},"laneRef":{"type":"integer","format":"int32"},"pos":{"type":"number","format":"double"},"type":{"type":"string","enum":["island","side","face_to_face","staggered"]},"parkingLots":{"type":"integer","format":"int32"},"address":{"type":"string"},"center":{"type":"string"},"line":{"$ref":"#/components/schemas/BusLineResponse"},"coordinates":{"$ref":"#/components/schemas/Coordinates"}}},"PublicTransitResponse":{"type":"object","properties":{"busStations":{"type":"array","items":{"$ref":"#/components/schemas/BusStationResponse"}}}},"RoadAssetData":{"type":"object","properties":{"pavementMarkings":{"type":"array","items":{"$ref":"#/components/schemas/PavementMarkingData"}}}},"PavementMarkingLogs":{"type":"object","properties":{"versionId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"data":{"$ref":"#/components/schemas/LogsData"}}},"CellResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"length":{"type":"number","format":"double"},"offset":{"type":"number","format":"double"}}},"ConnectionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fromLink":{"type":"integer","format":"int64"},"fromLane":{"type":"integer","format":"int64"},"fromLaneCoordinates":{"$ref":"#/components/schemas/Coordinates"},"toLink":{"type":"integer","format":"int64"},"toLane":{"type":"integer","format":"int64"},"toLaneCoordinates":{"$ref":"#/components/schemas/Coordinates"},"turning":{"type":"string","enum":["Straight","Right_Turn","Left_Turn"]},"length":{"type":"number","format":"double"},"width":{"type":"number","format":"double"},"ffSpd":{"type":"number","format":"double"},"shape":{"type":"string"},"coordinates":{"type":"array","items":{"$ref":"#/components/schemas/Coordinates"}}}},"LaneResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"leftLaneId":{"type":"string"},"rightLaneId":{"type":"string"},"numCell":{"type":"integer","format":"int32"},"laneAccessType":{"type":"string"},"rightLC":{"type":"boolean"},"leftLC":{"type":"boolean"},"shape":{"type":"string"},"segments":{"type":"array","items":{"$ref":"#/components/schemas/SegmentResponse"}},"cells":{"type":"array","items":{"$ref":"#/components/schemas/CellResponse"}},"coordinates":{"type":"array","items":{"$ref":"#/components/schemas/Coordinates"}}}},"LinkResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fromNode":{"type":"integer","format":"int64"},"toNode":{"type":"integer","format":"int64"},"numLane":{"type":"integer","format":"int32"},"length":{"type":"number","format":"double"},"width":{"type":"number","format":"double"},"maxSpd":{"type":"number","format":"double"},"minSpd":{"type":"number","format":"double"},"ffSpd":{"type":"number","format":"double"},"waveSpd":{"type":"number","format":"double"},"qmax":{"type":"number","format":"double"},"maxVeh":{"type":"number","format":"double"},"simType":{"type":"string","enum":["Meso","Micro"]},"type":{"type":"string","enum":["straight","curve"]},"layer":{"type":"string"},"stopLine":{"type":"number","format":"double"},"shape":{"type":"string"},"lanes":{"type":"array","items":{"$ref":"#/components/schemas/LaneResponse"}},"sections":{"type":"array","items":{"$ref":"#/components/schemas/SectionResponse"}},"coordinates":{"type":"array","items":{"$ref":"#/components/schemas/Coordinates"}}}},"NetworkResponse":{"type":"object","properties":{"name":{"type":"string"},"id":{"type":"integer","format":"int64"},"nodes":{"type":"array","items":{"$ref":"#/components/schemas/NodeResponse"}},"links":{"type":"array","items":{"$ref":"#/components/schemas/LinkResponse"}}}},"NodeResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["Intersection","Normal","Merging","Diverging","Terminal","Garage"]},"numPort":{"type":"integer","format":"int32"},"numConnection":{"type":"integer","format":"int32"},"v2x":{"type":"string","enum":["on","off"]},"center":{"type":"string"},"ports":{"type":"array","items":{"$ref":"#/components/schemas/PortResponse"}},"connections":{"type":"array","items":{"$ref":"#/components/schemas/ConnectionResponse"}},"coordinates":{"$ref":"#/components/schemas/Coordinates"},"portLinkIds":{"type":"array","items":{"type":"string"}}}},"PortResponse":{"type":"object","properties":{"type":{"type":"string","enum":["in","out"]},"linkId":{"type":"string"},"direction":{"type":"integer","format":"int64"}}},"SectionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"leftId":{"type":"string"},"rightId":{"type":"string"},"slope":{"type":"number","format":"double"},"length":{"type":"number","format":"double"},"offset":{"type":"number","format":"double"}}},"SegmentResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"block":{"type":"boolean"},"initPoint":{"type":"number","format":"double"},"endPoint":{"type":"number","format":"double"}}},"MenuTreeResponse":{"type":"object","properties":{"menuId":{"type":"integer","format":"int64"},"menuCode":{"type":"string"},"language":{"type":"string"},"nameKor":{"type":"string"},"nameEn":{"type":"string"},"depth":{"type":"integer","format":"int32"},"sortOrder":{"type":"integer","format":"int32"},"available":{"type":"string"},"accessRole":{"type":"string","enum":["ROLE_ADMIN","ROLE_USER"]},"rootId":{"type":"integer","format":"int64"}}},"LayerResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"key":{"type":"string"},"label":{"type":"string"},"basic":{"type":"boolean"},"auth":{"type":"integer","format":"int32"},"formType":{"type":"string"},"url":{"type":"string"},"groupKey":{"type":"string"}}},"LayerGroupResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"key":{"type":"string"},"label":{"type":"string"},"layers":{"type":"array","items":{"$ref":"#/components/schemas/LayerResponse"}}}}}}}