# iitp-rest/src/main/java/com/iitp/iitp_rest/model/publicTransit/rail/TimetableResponse.java
package com.iitp.iitp_rest.model.publicTransit.rail;

import lombok.Data;

import java.util.List;

@Data
public class TimetableResponse {
    private String dayOfWeek;
    private String lineId;
    private List<String> times;
}